move to the /pages folder
