+++ meshapp/src/App.tsx
@@ -1,7 +1,7 @@
 import "./App.css";
 import AppRoutes from "./routing/AppRoutes";
-import ProfilePage from "./profile/profile-page";
-import { exampleProfile } from "./profile/tests/profile-examples";
+import ProfilePage from "./pages/profile/profile-page";
+import { exampleProfile } from "./pages/profile/tests/profile-examples";
 import { ThemeContextProvider } from "./themes/ThemeContextProvider";
 import { MainThemeProvider } from "./themes/MainThemeProvider";
 import { AccountProvider } from "./contexts/UserContext";
