+++ meshapp/src/pages/profile/styling/profile-page.css
@@ -2,11 +2,6 @@ body {
   background-color: #0b7d66;
 }
 
-@font-face {
-  font-family: "cocogoose";
-  src: url("../../../public/fonts/cocogoose.ttf");
-}
-
 .profile-page-container {
   position: relative;
   margin: 80px 175px;
